+++ CHANGELOG.md
@@ -14,6 +14,7 @@ CHANGELOG
 ---
 
  * Deprecate using `Request::sendHeaders()` after headers have already been sent; use a `StreamedResponse` instead
+ * Add support for the `QUERY` HTTP method
 
 7.3
 ---